Update components with styled components 🚀🚀
- Style components with styled components 🖌
- Remove redundancy 🚮
- Enforce consistency ⚒️
- Make the world a better place 🚀
