+++ src/component/menu/menucomponents/popup/popup.styled.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components"
+
+export const PopupWrap = styled.div`
+    position: fixed;
+    z-index: 10;
+`
+
+export const PopupCover = styled.div`
+    position: fixed;
+    top: -100vh;
+    left: -100vw;
+    height: 200vh;
+    width: 200vw;
+`
