+++ src/component/menu/menucomponents/stylepicker/shapetools/shapetools.styled.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components"
+
+export const StyledShapeTools = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    border-radius: var(--menubar-border-radius);
+    background: var(--menubar-background);
+    width: var(--button-width);
+`
