+++ src/component/menu/menucomponents/stylepicker/widthpicker/widthpicker.styled.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components"
+
+export const WidthTools = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    border-radius: var(--menubar-border-radius);
+    background: var(--menubar-background);
+    width: var(--button-width);
+`
+
+export const WidthToolsSlider = styled.div`
+    padding-top: 4px;
+    padding-bottom: 4px;
+    height: 100%;
+`
