+++ src/css/menubars.css
@@ -36,34 +36,3 @@
     left: 0px;
     box-shadow: var(--menubar-box-shadow);
 }
-
-.favtoolbox {
-    position: relative;
-    height: var(--button-height);
-}
-
-.favtool-width {
-    background: rgba(0, 0, 0, 0.733);
-    border-radius: var(--button-border-radius);
-    height: 13px;
-    width: 13px;
-    position: absolute;
-    bottom: -3px;
-    right: -3px;
-    color: white;
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-        Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-    text-align: center;
-    font-size: 12px;
-}
-
-.favtooloptions {
-    position: absolute;
-    bottom: 3px; /* same as above bottom */
-    left: 45px;
-    display: grid;
-    grid-auto-flow: column;
-    border-radius: var(--menubar-border-radius);
-    background-color: var(--menubar-background);
-    box-shadow: var(--menubar-box-shadow);
-}
