--- src/css/menucomponents/pageoptions.css
@@ -1,47 +1,0 @@
-.pageoptions-wrap {
-    display: grid;
-    height: var(--button-height);
-    grid-auto-flow: column;
-    gap: var(--button-gap);
-}
-
-.pageoptions {
-    padding-top: 2px;
-    padding-bottom: 2px;
-    padding-left: 5px;
-    padding-right: 5px;
-    position: absolute;
-    top: 40px;
-    left: -5px;
-    display: grid;
-    grid-auto-flow: column;
-    border-radius: var(--menubar-border-radius);
-    background-color: var(--menubar-background);
-}
-
-.page-blank,
-.page-checkered,
-.page-ruled {
-    height: 50px;
-    width: 50px;
-    outline: none;
-    border-radius: 5px;
-    border-width: 1px;
-}
-
-.page-blank {
-    background: white;
-}
-
-.page-checkered {
-    background-image: repeating-linear-gradient(90deg, transparent 0px, transparent 9px, teal 10px),
-        repeating-linear-gradient(0deg, white 0px, white 9px, teal 10px);
-}
-
-.page-ruled {
-    background-image: repeating-linear-gradient(
-        white 0px,
-        white 9px,
-        teal 10px
-    );
-}
