--- src/css/menucomponents/popup.css
@@ -1,18 +1,0 @@
-.popup {
-    position: fixed;
-    z-index: 10;
-}
-
-.pagesettingspopup {
-    position: relative;
-    top: 0px;
-    z-index: 10;
-}
-
-.cover {
-    position: fixed;
-    top: -100vh;
-    left: -100vw;
-    height: 200vh;
-    width: 200vw;
-}
