--- src/css/sessioninfo.css
@@ -1,44 +1,0 @@
-.session-info {
-    position: fixed;
-    left: 10px;
-    top: 90px;
-    overflow: hidden;
-    border-radius: var(--menubar-border-radius);
-    max-height: 500px;
-    pointer-events: none;
-    display: grid;
-    grid-auto-flow: row;
-    gap: 10px;
-}
-
-.user-info {
-    background: #00000066;
-    display: grid;
-    grid-auto-flow: column;
-    font-family: 'Courier New', Courier, monospace;
-    font-size: small;
-    height: 18px;
-    width: fit-content;
-    max-width: 250px;
-    gap: 5px;
-    border-radius: 100px;
-    padding-left: 1.5px;
-    padding-right: 6px;
-}
-
-.user-color {
-    background: #00000088;
-    width: 15px;
-    height: 15px;
-    border-radius: 100px;
-    margin: auto;
-}
-
-.user-alias {
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    color: white;
-    width: 100%;
-    margin: auto;
-}
