--- src/css/stylepicker.css
@@ -1,43 +1,0 @@
-.style-picker {
-    display: flex;
-    flex-direction: row;
-    background: #00000088;
-    border-radius: var(--menubar-border-radius);
-    height: 200px;
-    width: 280px;
-    position: fixed;
-    top: 50px;
-    left: 50%;
-    transform: translateX(-50%);
-    /* position: absolute;
-    top: 10px;
-    left: -5px; */
-    padding: 5px;
-    box-shadow: var(--box-shadow);
-    justify-content: space-between;
-}
-
-.color-picker .react-colorful__saturation-pointer,
-.color-picker .react-colorful__hue-pointer,
-.color-picker .react-colorful__alpha-pointer {
-    width: 14px;
-    height: 14px;
-    border-radius: 50px;
-}
-
-.shapetools,
-.width-tools {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: center;
-    border-radius: var(--menubar-border-radius);
-    background: var(--menubar-background);
-    width: var(--button-width);
-}
-
-.width-slider-wrap {
-    padding-top: 4px;
-    padding-bottom: 4px;
-    height: 100%;
-}
