+++ src/view/whiteboard.tsx
@@ -2,13 +2,13 @@ import React, { useEffect } from "react"
 import { useParams } from "react-router-dom"
 import { handleAddPageOver } from "../drawing/handlers"
 import boardKeyListener from "../component/board/keylistener"
-import Toolbar from "../component/menu/toolbar"
+import Toolbar from "../component/menu/toolbar/toolbar"
 import BoardStage from "../component/board/stage"
-import SessionInfo from "../component/menu/menucomponents/sessioninfo"
-import ViewNav from "../component/menu/viewnavigation"
+import SessionInfo from "../component/menu/menucomponents/sessioninfo/sessioninfo"
+import ViewNav from "../component/menu/viewnavigation/viewnavigation"
 import { SET_SDIAG } from "../redux/slice/webcontrol"
 import { isConnected, pingSession } from "../api/websocket"
-import FavTools from "../component/menu/favtools"
+import FavTools from "../component/menu/favtools/favtools"
 import { useCustomDispatch } from "../redux/hooks"
 
 const Whiteboard: React.FC = () => {
